services/base/AutoDiscServer/libs/auto_disc/pyproject.toml
[project]
name = "auto_disc"
version = "0.1.0"
description = "Pure Python library for automated, curiosity-based exploration."
readme = "README.md"
requires-python = ">=3.6"
license = {file = "LICENSE"}
authors = [
  {name = "Jesse Young Lin", email = "jlin@inria.fr"},
  {name = "Clément Romac", email = "cromac@inria.fr"},
]
maintainers = [
  {name = "Jesse Young Lin", email = "jlin@inria.fr"}
]

dependencies = [
  "addict==2.4.0",
  "matplotlib==3.5.3",
  "numpy==1.21.6",
  "pillow==9.2.0",
  "requests==2.28.1",
  "graphviz==0.20.1",
  "neat-python==0.92",
  "torch==1.7.1",
  "tinydb==4.7.0",
  "imageio==2.9.0",
  "imageio-ffmpeg==0.4.7",
  "sqlalchemy==2.0.3"
]

[project.optional-dependencies]
test = ["pytest", "pytest-mock", "coverage"]
